+++ feature-dapp-impl/src/main/java/io/novafoundation/nova/feature_dapp_impl/presentation/favorites/DAppFavoritesViewModel.kt
@@ -5,12 +5,12 @@ import io.novafoundation.nova.common.mixin.actionAwaitable.ActionAwaitableMixin
 import io.novafoundation.nova.common.mixin.actionAwaitable.confirmingAction
 import io.novafoundation.nova.feature_dapp_api.presentation.browser.main.DAppBrowserPayload
 import io.novafoundation.nova.feature_dapp_impl.data.model.FavouriteDApp
-import io.novafoundation.nova.feature_dapp_impl.data.model.mapDAppModelFavorite
-import io.novafoundation.nova.feature_dapp_impl.data.model.mapFavoriteDappToDappModel
 import io.novafoundation.nova.feature_dapp_impl.domain.DappInteractor
 import io.novafoundation.nova.feature_dapp_impl.presentation.DAppRouter
 import io.novafoundation.nova.feature_dapp_impl.presentation.common.DappModel
 import io.novafoundation.nova.feature_dapp_impl.presentation.common.favourites.RemoveFavouritesPayload
+import io.novafoundation.nova.feature_dapp_impl.presentation.common.mapDAppModelToFavorite
+import io.novafoundation.nova.feature_dapp_impl.presentation.common.mapFavoriteDappToDappModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
@@ -50,7 +50,7 @@ class DAppFavoritesViewModel(
 
     fun changeDAppOrdering(newOrdering: List<DappModel>) = launch {
         val favoriteItems = newOrdering.mapIndexed { index, dappModel ->
-            mapDAppModelFavorite(dappModel, index)
+            mapDAppModelToFavorite(dappModel, index)
         }
 
         interactor.updateFavoriteDapps(favoriteItems)
